+++ satella/__init__.py
@@ -1,2 +1,2 @@
 # coding=UTF-8
-__version__ = '2.2.2a1'
+__version__ = '2.2.2a2'