+++ satella/exception_handling/__init__.py
@@ -1,10 +1,12 @@
 from .dump_to_file import *
 from .exception_handlers import *
 from .global_eh import *
+from .memerrhandler import MemoryErrorExceptionHandler
 
 __all__ = [
     'GlobalExcepthook',
     'BaseExceptionHandler', 'exception_handler', 'FunctionExceptionHandler',
     'NORMAL_PRIORITY', 'ALWAYS_LAST', 'ALWAYS_FIRST',
+    'MemoryErrorExceptionHandler',
     'DumpToFileHandler', 'AsStream'
 ]