+++ satella/exception_handling/exception_handlers.py
@@ -1,5 +1,5 @@
-import types
 import typing as tp
+import types
 from abc import abstractmethod
 
 __all__ = [
@@ -8,7 +8,8 @@ __all__ = [
     'exception_handler',
     'ALWAYS_FIRST',
     'ALWAYS_LAST',
-    'NORMAL_PRIORITY'
+    'NORMAL_PRIORITY',
+    'ExceptionHandlerCallable'
 ]
 
 ALWAYS_FIRST = -1000
@@ -16,28 +17,50 @@ NORMAL_PRIORITY = 0
 ALWAYS_LAST = 1000
 
 
+ExceptionHandlerCallable = tp.Callable[[type, BaseException, types.TracebackType], bool]
+
+
 class BaseExceptionHandler:
     def __init__(self, priority=NORMAL_PRIORITY):
+        """
+        Instantiate an exception handler with provided priority.
+        Handlers with smaller priorities run sooner.
+
+        :param priority: Priority to use for this handler
+        """
         self.priority = priority
 
-    def install(self):
+    def install(self) -> 'BaseExceptionHandler':
+        """
+        Register this handler to run upon exceptions
+        """
         from .global_eh import GlobalExcepthook
         GlobalExcepthook().add_hook(self)
         return self
 
     def uninstall(self):
+        """
+        Unregister this handler to run on exceptions
+        """
         from .global_eh import GlobalExcepthook
         GlobalExcepthook().remove_hook(self)
 
     @abstractmethod
-    def handle_exception(self, type_: tp.Optional[type], value,
+    def handle_exception(self, type_: tp.Callable[[type, BaseException, types.TracebackType], None], value,
                          traceback: types.TracebackType) -> tp.Optional[bool]:
-        """Return True to intercept the exception. It won't be propagated to other handlers."""
+        """
+        Return True to intercept the exception, so that it won't be propagated to other handlers.
+        """
         pass
 
 
 class FunctionExceptionHandler(BaseExceptionHandler):
-    def __init__(self, fun: tp.Callable, priority: int = NORMAL_PRIORITY):
+    """
+    A exception handler to make callables of given signature into Satella's exception handlers.
+
+    Your exception handler must return a bool, whether to intercept the exception and not propagate it.
+    """
+    def __init__(self, fun: ExceptionHandlerCallable, priority: int = NORMAL_PRIORITY):
         super(FunctionExceptionHandler, self).__init__(priority)
         self.fun = fun
 
@@ -52,7 +75,7 @@ def exception_handler(priority: int = NORMAL_PRIORITY):
     Convert a callable to an FunctionExceptionHandler. Usage
 
     >>> @exception_handler(priority=-10)
-    >>> def handle_exc(type, val, traceback):
+    >>> def handle_exc(type_, val, traceback):
     >>>     ...
 
     :return: ExceptionHandler instance
@@ -61,7 +84,7 @@ def exception_handler(priority: int = NORMAL_PRIORITY):
     if not isinstance(priority, int):
         raise TypeError('Did you forget to use it as @exception_handler() ?')
 
-    def outer(fun):
+    def outer(fun: ExceptionHandlerCallable) -> FunctionExceptionHandler:
         return FunctionExceptionHandler(fun, priority=priority)
 
     return outer