+++ satella/posix/pidlock.py
@@ -1,109 +1,95 @@
 import logging
 import os
 
+import psutil
+
 logger = logging.getLogger(__name__)
 
 
-class FailedToAcquire(Exception):
-    """Failed to acquire the process lock file"""
-
-
-class LockIsHeld(FailedToAcquire):
+class LockIsHeld(Exception):
     """
     Lock is held by someone
 
-    Has two attributes:
-        pid - integer - PID of the holder
-        is_alive - bool - whether the holder is an alive process
+    pid -- PID of the holder, who is alive
     """
 
+    def __init__(self, pid):
+        self.pid = pid
 
-class AcquirePIDLock:
+
+class PIDFileLock:
     """
     Acquire a PID lock file.
 
+    Usable also on Windows
+
     Usage:
 
-    >>> with AcquirePIDLock('myservice.pid'):
+    >>> with PIDFileLock('myservice.pid'):
     >>>     ... rest of code ..
 
     Or alternatively
 
-    >>> pid_lock = AcquirePIDLock('myservice.pid')
+    >>> pid_lock = PIDFileLock('myservice.pid')
     >>> pid_lock.acquire()
     >>> ...
     >>> pid_lock.release()
 
     The constructor doesn't throw, __enter__ or acquire() does, one of:
 
-    * AcquirePIDLock.FailedToAcquire - base class for errors. Thrown if can't read the file
-    * AcquirePIDLock.LockIsHeld - lock is already held. This has two attributes - pid (int), the PID of holder,
+   * LockIsHeld - lock is already held. This has two attributes - pid (int), the PID of holder,
                                   and is_alive (bool) - whether the holder is an alive process
     """
 
-    def __init__(self, pid, is_alive):
-        self.pid = pid
-        self.is_alive = is_alive
-
-    def __init__(self, pid_file, base_dir=u'/var/run', delete_on_dead=False):
+    def __init__(self, pid_file, base_dir=u'/var/run'):
         """
         Initialize a PID lock file object
 
         :param pid_file: rest of path
         :param base_dir: base lock directory
-        :param delete_on_dead: delete the lock file if holder is dead, and retry
         """
-        self.delete_on_dead = delete_on_dead
-
         self.path = os.path.join(base_dir, pid_file)
-
-        self.fileno = None
+        self.file_no = None
 
     def release(self):
-        if self.fileno is not None:
-            os.close(self.fileno)
+        """
+        Free the lock
+        """
+        if self.file_no is not None:
             os.unlink(self.path)
+            self.file_no = None
 
     def acquire(self):
         """
         Acquire the PID lock
 
         :raises LockIsHeld: if lock if held
-        :raises FailedToAcquire: if for example a directory exists in that place
         """
         try:
-            self.fileno = os.open(self.path, os.O_CREAT | os.O_EXCL)
-        except (IOError, OSError):
-            try:
-                with open(self.path, 'rb') as flock:
-                    try:
-                        pid = int(flock.read())
-                    except ValueError:
-                        logger.warning(
-                            'PID file found but doesn''t have an int, skipping')
-                        return
-            except IOError as e:
-                raise FailedToAcquire(repr(e))
+            self.file_no = os.open(self.path, os.O_CREAT | os.O_EXCL | os.O_WRONLY)
+        except (OSError, FileExistsError):
+            with open(self.path, 'r') as fin:
+                data = fin.read().strip()
 
-            # Is this process alive?
             try:
-                os.kill(pid, 0)
-            except OSError:  # dead
-                raise LockIsHeld(pid, False)
+                pid = int(data)
+            except ValueError:
+                os.unlink(self.path)
+                return self.acquire()
+
+            if pid in {x.pid for x in psutil.process_iter()}:
+                raise LockIsHeld(pid)
             else:
-                raise LockIsHeld(pid, True)
+                # does not exist
+                os.unlink(self.path)
+                return self.acquire()
+
+        fd = os.fdopen(self.file_no, 'w')
+        fd.write(str(os.getpid()) + '\n')
+        fd.close()
 
     def __enter__(self):
-        try:
-            self.acquire()
-        except LockIsHeld as e:
-            if self.delete_on_dead and (not e.is_alive):
-                os.unlink(self.path)
-                self.acquire()
-            else:
-                raise
-
-        self.success = True
+        self.acquire()
 
     def __exit__(self, exc_type, exc_val, exc_tb):
         self.release()