+++ tests/test_configuration/test_sources/test_envvars.py
@@ -1,3 +1,4 @@
+from satella.configuration import sources
 from satella.configuration.sources import EnvVarsSource, OptionalSource, AlternativeSource, EnvironmentSource, \
     StaticSource, MergingSource
 from .utils import SourceTestCase, mock_env
@@ -6,7 +7,7 @@ from .utils import SourceTestCase, mock_env
 class TestEnvVarsSource(SourceTestCase):
     @mock_env('satella', '{"a":2}')
     def test_ok(self):
-        self.assertSourceHas(EnvVarsSource('satella'), {u"a": 2})
+        self.assertSourceHas(sources.EnvVarsSource('satella'), {u"a": 2})
 
     def test_none(self):
         self.assertSourceEmpty(OptionalSource(EnvVarsSource('satella')))