+++ satella/__init__.py
@@ -1,2 +1,2 @@
 # coding=UTF-8
-__version__ = '2.2.3'
+__version__ = '2.2.4a1'