+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 # v2.2.2
 
-* _TBA_
+* more docs
+* fixed `MemoryErrorExceptionHandler` not being importable from `satella.exception_handling`
+* fixed the custom_hook for `MemoryErrorExceptionHandler`: it won't
+  kill everything if the custom_hook returns True 
 
 # v2.2.1